requirements.txt
cffi==1.14.4
cryptography==3.3.1
Django==2.2.3
django-crispy-forms==1.11.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.11.0
mysql-connector==2.2.9
mysqlclient==1.4.6
pycparser==2.20
PyMySQL==0.9.2
pyodbc==4.0.30
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
